james' validate() function (obsolete?)
I wrote a validate() function last night and thought it had synced, but it hadn't. This morning Bryson updated last3digits() to validate input by itself. I've commented my validate() function out but left it in in case we can find a use for it. 

I'll be including the isdigit() loop in last3digits() to better validate user input when they enter the three characters. For this purpose, I've updated main to #include <ctype.h>.